+++ backend/app/api/v1/system/user/model.py
@@ -2,7 +2,7 @@
 用户数据模型
 """
 
-from sqlalchemy import Column, String, Integer, DateTime, Text, JSON, Table, ForeignKey
+from sqlalchemy import Column, String, Integer, BigInteger, DateTime, Text, JSON, Table, ForeignKey
 from sqlalchemy.orm import relationship
 from app.core.base_model import BaseModel
 
@@ -10,8 +10,8 @@ from app.core.base_model import BaseModel
 user_role = Table(
     'sys_user_role',
     BaseModel.metadata,
-    Column('user_id', Integer, ForeignKey('sys_user.id', ondelete='CASCADE'), primary_key=True),
-    Column('role_id', Integer, ForeignKey('sys_role.id', ondelete='CASCADE'), primary_key=True),
+    Column('user_id', BigInteger, ForeignKey('sys_user.id', ondelete='CASCADE'), primary_key=True),
+    Column('role_id', BigInteger, ForeignKey('sys_role.id', ondelete='CASCADE'), primary_key=True),
 )
 
 
@@ -32,7 +32,7 @@ class UserModel(BaseModel):
     status = Column(Integer, nullable=False, default=1, comment="用户状态（0:禁用 1:启用）")
     gender = Column(Integer, nullable=True, default=0, comment="性别（0:未知 1:男 2:女）")
     
-    dept_id = Column(Integer, ForeignKey('sys_dept.id'), nullable=True, comment="部门ID")
+    dept_id = Column(BigInteger, ForeignKey('sys_dept.id'), nullable=True, comment="部门ID")
     post = Column(String(100), nullable=True, comment="岗位")
     remark = Column(String(500), nullable=True, comment="备注")
     tags = Column(JSON, nullable=True, comment="个性标签（JSON数组）")
